Working toward 3D
Turtle factored so that the movement code (string interpretation) is in the base.
Added an obj turtle that renders out lines as cubes in an wavefront obj file.
Added some tests for the exe for an annoying utility function I needed to make.
Added some render obj calls in main:Do3D code.